test(graph): add graph provenance contract suites
